+++ src/log/mod.rs
@@ -25,7 +25,7 @@ mod port;
 /// # [memory]
 /// # flash = { size = \"128K\", origin = 0x08000000 }
 /// # ram = { size = \"20K\", origin = 0x20000000 }
-/// # [heap]
+/// # [heap.main]
 /// # size = \"0\"
 /// # pools = []
 /// # [linker]